Move YlmSpherepack to NumericalAlgorithms/SphericalHarmonics.
This should fix #3573, and it does some of #3381
