+++ src/Domain/CoordinateMaps/TimeDependent/Shape.cpp
@@ -12,12 +12,11 @@
 #include <pup_stl.h>
 #include <string>
 
-#include "ApparentHorizons/SpherepackIterator.hpp"
-#include "ApparentHorizons/TagsTypeAliases.hpp"
 #include "DataStructures/DataVector.hpp"
 #include "DataStructures/Tensor/EagerMath/DeterminantAndInverse.hpp"
 #include "Domain/CoordinateMaps/TimeDependent/ShapeMapTransitionFunctions/ShapeMapTransitionFunction.hpp"
 #include "Domain/FunctionsOfTime/FunctionOfTime.hpp"
+#include "NumericalAlgorithms/SphericalHarmonics/SpherepackIterator.hpp"
 #include "Utilities/ContainerHelpers.hpp"
 #include "Utilities/DereferenceWrapper.hpp"
 #include "Utilities/ErrorHandling/Assert.hpp"
