+++ src/NumericalAlgorithms/CMakeLists.txt
@@ -13,3 +13,4 @@ add_subdirectory(OdeIntegration)
 add_subdirectory(RootFinding)
 add_subdirectory(SpatialDiscretization)
 add_subdirectory(Spectral)
+add_subdirectory(SphericalHarmonics)
