+++ tests/Unit/Helpers/NumericalAlgorithms/SphericalHarmonics/CMakeLists.txt
@@ -1,10 +1,10 @@
 # Distributed under the MIT License.
 # See LICENSE.txt for details.
 
-set(LIBRARY "ApparentHorizonsHelpers")
+set(LIBRARY "SphericalHarmonicsHelpers")
 
 set(LIBRARY_SOURCES
-  StrahlkorperGrTestHelpers.cpp
+  YlmTestFunctions.cpp
   )
 
 add_spectre_library(${LIBRARY} ${LIBRARY_SOURCES})
@@ -12,7 +12,6 @@ add_spectre_library(${LIBRARY} ${LIBRARY_SOURCES})
 target_link_libraries(
   ${LIBRARY}
   PUBLIC
-  ApparentHorizons
   DataStructures
   SphericalHarmonics
   Utilities
