+++ tests/Unit/NumericalAlgorithms/CMakeLists.txt
@@ -12,3 +12,4 @@ add_subdirectory(LinearSolver)
 add_subdirectory(OdeIntegration)
 add_subdirectory(RootFinding)
 add_subdirectory(Spectral)
+add_subdirectory(SphericalHarmonics)
